# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "BPMT API's",
        "description": "Project management operations",
        "version": "1.0.0"
    },
    "host": "http://my-default-host.com",
    "basePath": "/api/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/v1/milestone": {
            "post": {
                "tags": [
                    "Milestones"
                ],
                "summary": "milestone creation",
                "description": "Returns created milestone",
                "operationId": "create-milestone",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "authorization header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "formData",
                        "description": "id of the milestone at the time of update",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "description": "title for milestone",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "description": "description of the milestone",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "startDate",
                        "in": "formData",
                        "description": "start datetime",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "endDate",
                        "in": "formData",
                        "description": "End Datetime",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "estimatedHours",
                        "in": "formData",
                        "description": "Estimated Hours",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "status",
                        "in": "formData",
                        "description": "Milestone Status('created', 'assigned', 'onhold', 'inprogress','completed', 'cancelled',' failed')",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "project_id",
                        "in": "formData",
                        "description": "Id of the project",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/v1/milestone/{id}": {
            "get": {
                "tags": [
                    "Milestones"
                ],
                "summary": "single Milestones details",
                "description": "Returns single Milestone details",
                "operationId": "single Milestone Details",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "authorization header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Milestone Id",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Milestones"
                ],
                "summary": "Delete a Milestone",
                "description": "Delete a Milestone",
                "operationId": "delete Milestone",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "authorization header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Milestone Id",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/v1/milestone/by-project/{id}": {
            "get": {
                "tags": [
                    "Milestones"
                ],
                "summary": "Milestone from projects",
                "description": "Returns Milestone from projects",
                "operationId": "All-Milestone-from-projects",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "authorization header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project Id",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    },
...
}